test: use dynamic port for web proxy tests
This changes the web proxy tests to use a dynamic, unused port instead
of the default 8080 port, which could cause the tests to fail if 8080
is being used.
