+++ test/integration/WebProxy.spec.ts
@@ -1,18 +1,20 @@
 import chai, { expect } from 'chai';
 import chaiHttp from 'chai-http';
 import Xud from '../../lib/Xud';
+import { getUnusedPort } from '../utils';
 
-describe('WebProxy', () => {
+describe('WebProxy', async () => {
   let xud: Xud;
   let config: any;
   chai.use(chaiHttp);
+  const port = await getUnusedPort();
 
   before(async () => {
     config = {
       dbpath: ':memory:',
       webproxy: {
+        port,
         disable: false,
-        port: 8080,
       },
       logpath: '',
       loglevel: 'warn',
