+++ test/utils.ts
@@ -1,0 +1,18 @@
+import net from 'net';
+
+/**
+ * Discovers and returns a dynamically assigned, unused port available for testing.
+ */
+export const getUnusedPort = async () => {
+  return new Promise<number>((resolve, reject) => {
+    const server = net.createServer();
+    server.unref();
+    server.on('error', reject);
+    server.listen(0, () => {
+      const { port } = server.address();
+      server.close(() => {
+        resolve(port);
+      });
+    });
+  });
+};
